requirements.txt
yfinance>=1.0.0
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0
matplotlib>=3.6.0
plotly>=5.11.0
scikit-learn>=1.1.0
PyYAML>=6.0
click>=8.1.0
